Move x86 hw into explicit module
